[fix] 자잘자잘 - [X] drawer들 위치 수정( friend-> cateDiary, feed -> friend) - [X] 스플래쉬 추가 - [X] colors 리소스 변경 및 추가
resolves: #22
