+++ app/src/main/AndroidManifest.xml
@@ -12,19 +12,23 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.Peering">
-        <activity
-            android:name=".activity.FeedWriteActivity" />
-        <activity android:name=".activity.LoginActivity"
-            android:exported="true" >
+        <activity android:name=".activity.SplashActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".activity.MainActivity"></activity>
-        <activity android:name="com.kakao.sdk.auth.AuthCodeHandlerActivity"
-            android:exported="true" >
+        <activity android:name=".activity.FeedWriteActivity" />
+        <activity
+            android:name=".activity.LoginActivity"
+            android:exported="true">
+
+        </activity>
+        <activity android:name=".activity.MainActivity" />
+        <activity
+            android:name="com.kakao.sdk.auth.AuthCodeHandlerActivity"
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.VIEW" />
