+++ app/src/main/java/com/awesomesol/peering/activity/FeedWriteActivity.kt
@@ -4,7 +4,7 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.awesomesol.peering.R
 import com.awesomesol.peering.databinding.ActivityFeedWriteBinding
-import com.awesomesol.peering.feed.FeedFragment
+import com.awesomesol.peering.friend.FeedFragment
 
 class FeedWriteActivity : AppCompatActivity() {
     private lateinit var binding : ActivityFeedWriteBinding
