+++ app/src/main/java/com/awesomesol/peering/activity/SplashActivity.kt
@@ -1,0 +1,29 @@
+package com.awesomesol.peering.activity
+
+import android.content.Intent
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import android.os.Handler
+import com.awesomesol.peering.R
+
+class SplashActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_splash)
+
+        Handler().postDelayed({
+            val intent = Intent(this, LoginActivity::class.java)
+            intent.addFlags(Intent.FLAG_ACTIVITY_NO_ANIMATION)
+            startActivity(intent)
+            finish()
+        },DURATION)
+
+    }
+    companion object {
+        private const val DURATION : Long = 3000
+    }
+
+    override fun onBackPressed() {
+        super.onBackPressed()
+    }
+}
