+++ app/src/main/java/com/awesomesol/peering/friend/FeedFragment.kt
@@ -1,7 +1,8 @@
-package com.awesomesol.peering.feed
+package com.awesomesol.peering.friend
 
 import android.content.Intent
 import android.os.Bundle
+import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
@@ -10,12 +11,26 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.awesomesol.peering.R
 import com.awesomesol.peering.activity.FeedWriteActivity
+import com.kakao.sdk.talk.TalkApiClient
 import kotlinx.android.synthetic.main.fragment_feed.*
 
 class FeedFragment : Fragment() {
-
+    val TAG="피드"
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+
+//        // 카카오톡 친구 목록 가져오기 (기본)
+//        TalkApiClient.instance.friends { friends, error ->
+//            if (error != null) {
+//                Log.e(TAG, "카카오톡 친구 목록 가져오기 실패", error)
+//            }
+//            else if (friends != null) {
+//                Log.i(TAG, "카카오톡 친구 목록 가져오기 성공 \n${friends.elements.joinToString("\n")}")
+//                Log.d(TAG, friends.toString())
+//                // 친구의 UUID 로 메시지 보내기 가능
+//            }
+//        }
+
 
     }
 
@@ -23,6 +38,8 @@ class FeedFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
+
+
         // Inflate the layout for this fragment
         val view = inflater.inflate(R.layout.fragment_feed, container, false)
         val rv = view.findViewById<RecyclerView>(R.id.rv_FeedFragment)
