+++ app/src/main/java/com/awesomesol/peering/friend/FeedRVAdapter.kt
@@ -1,9 +1,8 @@
-package com.awesomesol.peering.feed
+package com.awesomesol.peering.friend
 
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.awesomesol.peering.R
