+++ app/src/main/res/layout/fragment_feed.xml
@@ -5,7 +5,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".feed.FeedFragment">
+    tools:context=".friend.FeedFragment">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/tb_FeedFragment"
