+++ app/src/main/res/menu/bottom_navigation_main.xml
@@ -6,12 +6,12 @@
         android:icon="@drawable/bottom_home"
         android:title="캘린더" />
     <item
-        android:id="@+id/action_license_info"
+        android:id="@+id/action_community"
         android:enabled="true"
         android:icon="@drawable/bottom_home"
-        android:title="피드" />
+        android:title="종합" />
     <item
-        android:id="@+id/action_community"
+        android:id="@+id/action_license_info"
         android:enabled="true"
         android:icon="@drawable/bottom_home"
         android:title="친구" />
