+++ app/src/main/res/values/colors.xml
@@ -1,10 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="purple_200">#FFBB86FC</color>
-    <color name="purple_500">#FF6200EE</color>
-    <color name="purple_700">#FF3700B3</color>
-    <color name="teal_200">#FF03DAC5</color>
-    <color name="teal_700">#FF018786</color>
+    <color name="theme_red">#FFB6B9</color>
+    <color name="theme_yellow">#FAE3D9</color>
+    <color name="theme_green">#BBDED6</color>
+    <color name="theme_blue">#8AC6D1</color>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
 </resources>
